# Conflicts:
#	src/main/java/net/minecraft/client/renderer/entity/RenderPlayer.java

+++ src/main/java/net/inceptioncloud/dragonfly/ui/renderer/RenderUtils.java
@@ -2,36 +2,28 @@ package net.inceptioncloud.dragonfly.ui.renderer;
 
 import kotlin.ranges.RangesKt;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.Gui;
-import net.minecraft.client.gui.ScaledResolution;
-import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.client.renderer.OpenGlHelper;
-import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.client.renderer.WorldRenderer;
+import net.minecraft.client.gui.*;
+import net.minecraft.client.renderer.*;
 import net.minecraft.client.renderer.entity.RenderManager;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.util.AxisAlignedBB;
-import net.minecraft.util.BlockPos;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.entity.*;
 import net.minecraft.util.Timer;
+import net.minecraft.util.*;
 
 import java.awt.*;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
-import static java.lang.Math.*;
+import static java.lang.Math.PI;
+import static java.lang.Math.cos;
+import static java.lang.Math.sin;
 import static org.lwjgl.opengl.GL11.*;
 
-public final class RenderUtils
-{
+public final class RenderUtils {
     private static final Map<Integer, Boolean> glCapMap = new HashMap<>();
 
     private static final Minecraft mc = Minecraft.getMinecraft();
 
-    public static void drawAxisAlignedBB (final AxisAlignedBB axisAlignedBB, final Color color)
-    {
+    public static void drawAxisAlignedBB(final AxisAlignedBB axisAlignedBB, final Color color) {
         glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
         glEnable(GL_BLEND);
         glLineWidth(2F);
@@ -48,16 +40,14 @@ public final class RenderUtils
         glDisable(GL_BLEND);
     }
 
-    public static void drawPlatform (final double y, final Color color, final double size)
-    {
+    public static void drawPlatform(final double y, final Color color, final double size) {
         final RenderManager renderManager = mc.getRenderManager();
         final double renderY = y - renderManager.renderPosY;
 
         drawAxisAlignedBB(new AxisAlignedBB(size, renderY + 0.02D, size, -size, renderY, -size), color);
     }
 
-    public static void drawPlatform (final Entity entity, final Color color)
-    {
+    public static void drawPlatform(final Entity entity, final Color color) {
         final RenderManager renderManager = mc.getRenderManager();
         final Timer timer = mc.timer;
 
@@ -66,14 +56,20 @@ public final class RenderUtils
         final double z = entity.lastTickPosZ + (entity.posZ - entity.lastTickPosZ) * timer.renderPartialTicks - renderManager.renderPosZ;
 
         final AxisAlignedBB axisAlignedBB = entity.getEntityBoundingBox()
-            .offset(-entity.posX, -entity.posY, -entity.posZ)
-            .offset(x, y, z);
+                .offset(-entity.posX, -entity.posY, -entity.posZ)
+                .offset(x, y, z);
 
-        drawAxisAlignedBB(new AxisAlignedBB(axisAlignedBB.minX, axisAlignedBB.maxY + 0.2, axisAlignedBB.minZ, axisAlignedBB.maxX, axisAlignedBB.maxY + 0.26, axisAlignedBB.maxZ), color);
+        drawAxisAlignedBB(new AxisAlignedBB(
+                axisAlignedBB.minX,
+                axisAlignedBB.maxY + 0.2,
+                axisAlignedBB.minZ,
+                axisAlignedBB.maxX,
+                axisAlignedBB.maxY + 0.26,
+                axisAlignedBB.maxZ
+        ), color);
     }
 
-    public static void drawFilledBox (final AxisAlignedBB axisAlignedBB)
-    {
+    public static void drawFilledBox(final AxisAlignedBB axisAlignedBB) {
         final Tessellator tessellator = Tessellator.getInstance();
         final WorldRenderer worldRenderer = tessellator.getWorldRenderer();
         worldRenderer.begin(7, DefaultVertexFormats.POSITION);
@@ -138,8 +134,7 @@ public final class RenderUtils
         tessellator.draw();
     }
 
-    public static void drawRect (final float x, final float y, final float x2, final float y2, final int color)
-    {
+    public static void drawRect(final float x, final float y, final float x2, final float y2, final int color) {
         glEnable(GL_BLEND);
         glDisable(GL_TEXTURE_2D);
         glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
@@ -158,13 +153,19 @@ public final class RenderUtils
         glDisable(GL_LINE_SMOOTH);
     }
 
-    public static void drawRect (final float x, final float y, final float x2, final float y2, final Color color)
-    {
+    public static void drawRect(final float x, final float y, final float x2, final float y2, final Color color) {
         drawRect(x, y, x2, y2, color.getRGB());
     }
 
-    public static void drawBorderedRect (final float x, final float y, final float x2, final float y2, final float width, final int color1, final int color2)
-    {
+    public static void drawBorderedRect(
+            final float x,
+            final float y,
+            final float x2,
+            final float y2,
+            final float width,
+            final int color1,
+            final int color2
+    ) {
         drawRect(x, y, x2, y2, color2);
         glEnable(GL_BLEND);
         glDisable(GL_TEXTURE_2D);
@@ -189,14 +190,12 @@ public final class RenderUtils
         glDisable(GL_LINE_SMOOTH);
     }
 
-    public static void drawLoadingCircle (float x, float y)
-    {
+    public static void drawLoadingCircle(float x, float y) {
         int rot = (int) ((System.nanoTime() / 5000000) % 360);
         drawCircle(x, y, 10, rot - 180, rot);
     }
 
-    public static void drawCircle (float x, float y, float radius, int start, int end)
-    {
+    public static void drawCircle(float x, float y, float radius, int start, int end) {
         GlStateManager.enableBlend();
         GlStateManager.disableTexture2D();
         GlStateManager.tryBlendFuncSeparate(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA, GL_ONE, GL_ZERO);
@@ -206,7 +205,7 @@ public final class RenderUtils
         glLineWidth(2F);
 
         glBegin(GL_LINE_STRIP);
-        for (float i = end ; i >= start ; i -= 4)
+        for (float i = end; i >= start; i -= 4)
             glVertex2f((float) (x + (cos(i * PI / 180) * (radius * 1.001F))), (float) (y + (sin(i * PI / 180) * (radius * 1.001F))));
 
         glEnd();
@@ -216,8 +215,7 @@ public final class RenderUtils
         GlStateManager.disableBlend();
     }
 
-    public static void drawFilledCircle (int xx, int yy, float radius, Color col)
-    {
+    public static void drawFilledCircle(int xx, int yy, float radius, Color col) {
         int sections = 50;
         double dAngle = 2 * Math.PI / sections;
         float x, y;
@@ -229,7 +227,7 @@ public final class RenderUtils
         glEnable(GL_LINE_SMOOTH);
         glBegin(GL_TRIANGLE_FAN);
 
-        for (int i = 0 ; i < sections ; i++) {
+        for (int i = 0; i < sections; i++) {
             x = (float) (radius * Math.sin((i * dAngle)));
             y = (float) (radius * Math.cos((i * dAngle)));
 
@@ -245,8 +243,7 @@ public final class RenderUtils
         glPopMatrix();
     }
 
-    public static void drawArc (float x, float y, int start, int end, int width, int height, Color color)
-    {
+    public static void drawArc(float x, float y, int start, int end, float width, float height, Color color) {
         int factor = 5;
 
         glScaled(1D / factor, 1D / factor, 1D / factor);
@@ -257,7 +254,7 @@ public final class RenderUtils
         glBegin(GL_POLYGON);
         glVertex2f(x * factor, y * factor);
 
-        for (float i = end ; i >= start ; i--) {
+        for (float i = end; i >= start; i--) {
             glVertex2f((float) (x + (cos(i * PI / 180) * (width * 1.001F))) * factor, (float) (y + (sin(i * PI / 180) * (height * 1.001F))) * factor);
         }
 
@@ -269,8 +266,11 @@ public final class RenderUtils
         glBegin(GL_POLYGON);
         glVertex2f(x * factor, y * factor);
 
-        for (float i = end ; i >= start ; i--) {
-            glVertex2f((float) (x + (cos(i * PI / 180) * (width - (1F/factor)) * 1.001F)) * factor, (float) (y + (sin(i * PI / 180) * (height - (1F/factor)) * 1.001F)) * factor);
+        for (float i = end; i >= start; i--) {
+            glVertex2f(
+                    (float) (x + (cos(i * PI / 180) * (width - (1F / factor)) * 1.001F)) * factor,
+                    (float) (y + (sin(i * PI / 180) * (height - (1F / factor)) * 1.001F)) * factor
+            );
         }
 
         glEnd();
@@ -303,8 +303,7 @@ public final class RenderUtils
         GlStateManager.disableBlend();
     }
 
-    public static void drawImage (ResourceLocation image, int x, int y, int width, int height)
-    {
+    public static void drawImage(ResourceLocation image, int x, int y, int width, int height) {
         glDisable(GL_DEPTH_TEST);
         glEnable(GL_BLEND);
         glDepthMask(false);
@@ -318,8 +317,7 @@ public final class RenderUtils
         glEnable(GL_DEPTH_TEST);
     }
 
-    public static void drawImage (ResourceLocation image, int x, int y, int width, int height, float red, float green, float blue, float alpha)
-    {
+    public static void drawImage(ResourceLocation image, int x, int y, int width, int height, float red, float green, float blue, float alpha) {
         glDisable(GL_DEPTH_TEST);
         glEnable(GL_BLEND);
         glDepthMask(false);
@@ -332,13 +330,11 @@ public final class RenderUtils
         glEnable(GL_DEPTH_TEST);
     }
 
-    public static void glColor (final Color color)
-    {
+    public static void glColor(final Color color) {
         glColor(color.getRGB());
     }
 
-    private static void glColor (final int hex)
-    {
+    private static void glColor(final int hex) {
         float alpha = (hex >> 24 & 0xFF) / 255F;
         float red = (hex >> 16 & 0xFF) / 255F;
         float green = (hex >> 8 & 0xFF) / 255F;
@@ -346,8 +342,7 @@ public final class RenderUtils
         GlStateManager.color(red, green, blue, alpha);
     }
 
-    public static Color hexColor (final int hex)
-    {
+    public static Color hexColor(final int hex) {
         float alpha = (hex >> 24 & 0xFF) / 255F;
         float red = (hex >> 16 & 0xFF) / 255F;
         float green = (hex >> 8 & 0xFF) / 255F;
@@ -356,8 +351,14 @@ public final class RenderUtils
         return new Color(red, green, blue, alpha);
     }
 
-    public static void draw2D (final EntityLivingBase entity, final double posX, final double posY, final double posZ, final int color, final int backgroundColor)
-    {
+    public static void draw2D(
+            final EntityLivingBase entity,
+            final double posX,
+            final double posY,
+            final double posZ,
+            final int color,
+            final int backgroundColor
+    ) {
         GlStateManager.pushMatrix();
         GlStateManager.translate(posX, posY, posZ);
         glNormal3f(0F, 0F, 0F);
@@ -395,8 +396,7 @@ public final class RenderUtils
         GlStateManager.popMatrix();
     }
 
-    public static void draw2D (final BlockPos blockPos, final int color, final int backgroundColor)
-    {
+    public static void draw2D(final BlockPos blockPos, final int color, final int backgroundColor) {
         final RenderManager renderManager = mc.getRenderManager();
 
         final double posX = (blockPos.getX() + 0.5) - renderManager.renderPosX;
@@ -440,8 +440,7 @@ public final class RenderUtils
         GlStateManager.popMatrix();
     }
 
-    public static void drawLine (final double x, final double y, final double x1, final double y1, final float width)
-    {
+    public static void drawLine(final double x, final double y, final double x1, final double y1, final float width) {
         glDisable(GL_TEXTURE_2D);
         glLineWidth(width);
         glBegin(GL_LINES);
@@ -451,56 +450,60 @@ public final class RenderUtils
         glEnable(GL_TEXTURE_2D);
     }
 
-    public static void makeScissorBox (final float x, final float y, final float x2, final float y2)
-    {
+    public static void makeScissorBox(final float x, final float y, final float x2, final float y2) {
         final ScaledResolution scaledResolution = new ScaledResolution(mc);
         final int factor = scaledResolution.getScaleFactor();
-        glScissor((int) (x * factor), (int) ((scaledResolution.getScaledHeight() - y2) * factor), (int) ((x2 - x) * factor), (int) ((y2 - y) * factor));
+        glScissor(
+                (int) (x * factor),
+                (int) ((scaledResolution.getScaledHeight() - y2) * factor),
+                (int) ((x2 - x) * factor),
+                (int) ((y2 - y) * factor)
+        );
     }
 
-    public static void resetCaps ()
-    {
+    public static void resetCaps() {
         glCapMap.forEach(RenderUtils::setGlState);
     }
 
-    public static void enableGlCap (final int cap)
-    {
+    public static void enableGlCap(final int cap) {
         setGlCap(cap, true);
     }
 
-    public static void enableGlCap (final int... caps)
-    {
+    public static void enableGlCap(final int... caps) {
         for (final int cap : caps)
             setGlCap(cap, true);
     }
 
-    public static void disableGlCap (final int cap)
-    {
+    public static void disableGlCap(final int cap) {
         setGlCap(cap, true);
     }
 
-    public static void disableGlCap (final int... caps)
-    {
+    public static void disableGlCap(final int... caps) {
         for (final int cap : caps)
             setGlCap(cap, false);
     }
 
-    public static void setGlCap (final int cap, final boolean state)
-    {
+    public static void setGlCap(final int cap, final boolean state) {
         glCapMap.put(cap, glGetBoolean(cap));
         setGlState(cap, state);
     }
 
-    public static void setGlState (final int cap, final boolean state)
-    {
+    public static void setGlState(final int cap, final boolean state) {
         if (state)
             glEnable(cap);
         else
             glDisable(cap);
     }
 
-    public static void drawTriangle (final double x1, final double y1, final double x2, final double y2, final double x3, final double y3, final int color)
-    {
+    public static void drawTriangle(
+            final double x1,
+            final double y1,
+            final double x2,
+            final double y2,
+            final double x3,
+            final double y3,
+            final int color
+    ) {
         glColor(color);
         glBegin(GL_TRIANGLES);
         glVertex2d(x1, y1);
